cleanup core engine
