--- .release-it.json
@@ -1,29 +1,0 @@
-{
-    "buildCommand": "yarn build && yarn pack",
-    "requireCleanWorkingDir": false,
-    "src": {
-        "commit": true,
-        "commitMessage": "Release %s",
-        "tag": true,
-        "tagName": "%s",
-        "tagAnnotation": "Release %s",
-        "pushRepo": "origin",
-        "beforeStartCommand": "rimraf ./*.tgz",
-        "beforeStageCommand": false,
-        "afterReleaseCommand": "rimraf ./*.tgz",
-        "githubAssets": [
-            "*.tgz"
-        ]
-    },
-    "npm": {
-        "publish": false,
-        "publishPath": "dist/",
-        "tag": "latest"
-    },
-    "github": {
-        "release": true,
-        "assets": "*.tgz",
-        "tokenRef": "GITHUB_RELEASE_IT_TOKEN",
-        "note": "Must set tokenRef as environment variable"
-    }
-}
