+++ TODO.md
@@ -2,3 +2,10 @@ TODO
 ====
 
 * [ ] Write an article for implementing custom kites extension
+
+Discover Mode
+=============
+
+* `discover: true` - Discover extensions in rootDirectory
+* `discover: 'path-string'` - Discover extentions in `path-string`
+* `discover: false` - Disable discover option
