+++ packages/core/src/engine/kites.ts
@@ -24,7 +24,7 @@ export type KitesReadyCallback = (kites: IKites) => void;
  */
 export interface IKitesOptions {
   [key: string]: any;
-  discover?: boolean;
+  discover?: boolean | string; // string for path discovery
   loadConfig?: boolean;
   rootDirectory?: string;
   appDirectory?: string;
@@ -209,6 +209,22 @@ export class KitesInstance extends EventEmitter implements IKites {
       this.options.discover = true;
     }
     return this;
+  }
+
+  /**
+   * Thiết lập giá trị cấu hình cho các extensions
+   * Example:
+   *      .set('express:static', './assets') -> kites.options.express.static = './assets'
+   * @param option
+   * @param value
+   */
+  set(option: string, value: string) {
+    const tokens = option.split(':');
+    if (tokens.length === 2) {
+        this.options[tokens[0]][tokens[1]] = value;
+    } else if (tokens.length === 1) {
+        this.options[tokens[0]] = value;
+    }
   }
 
   /**
